==> deliverables/code/06_Kaggle_Submissions.ipynb <==
# Kaggle Submission

Anthony Amadasun

December 15th 20223

---

### Introduction

This section contains the code submission for the Kaggle competition. The purpose of this section is to practice refining models over time, utilizing train-test split, cross-validation, and data with unknown values for the target to simulate the modeling process. Moreover, familiarize myself with the use of Kaggle as a place to practice data science. 

The objective is to predict the sales price for each house. For each Id in the test set, I must predict the value of the SalePrice variable. 



---

### Importing Libraries

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from sklearn.linear_model import LinearRegression, Ridge, Lasso, RidgeCV
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.preprocessing import StandardScaler, PolynomialFeatures

---

### Load in Data

# load preprocessed data
df_cleaned = pd.read_csv('../data/test_clean4.csv')
file_path_train = '../data/train.csv'
df_train = pd.read_csv(file_path_train)

# df = pd.read_csv('https://raw.git.generalassemb.ly/dsi-1113/project-2/main/datasets/train.csv?token=AAAETZWEPZZRPXEMANF2EADFPIRVE')
# kaggle_data = pd.read_csv('https://raw.git.generalassemb.ly/dsi-1113/project-2/main/datasets/test.csv?token=AAAETZWYQX6PO7IXVWSDYQLFPIRXA')

---

### Create the Kaggle Model

# All numeric columns of the dataset selected
X = df_cleaned.select_dtypes(include=['float64', 'int64']).drop('SalePrice', axis=1)
y = df_cleaned['SalePrice']


X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=0.2, 
                                                    random_state=42)

# Scale our data.
scaler = StandardScaler()

# Fit/transform from training to learn mean, stdev. 
#and then transform both using things learned from training
Z_train = scaler.fit_transform(X_train)
Z_test = scaler.transform(X_test)

# hyperparameter
kaggle_alpha = 100

#Ridge model with specified alpha
kaggle_model = Ridge(alpha=kaggle_alpha)

# Fit the Ridge model
kaggle_model.fit(Z_train, y_train)

# Predictions for Ridge
y_pred_kaggle_train = kaggle_model.predict(Z_train)
y_pred_kaggle_test = kaggle_model.predict(Z_test)

# # Save Kaggle predictions to CSV
kaggle_predictions = pd.DataFrame({'SalePrice': y_pred_kaggle_test}, index=X_test.index)
# ridge_predictions.to_csv('../data/submission_ridge.csv')

# Ridge csv
kaggle_predictions.to_csv('../data/submission_kaggle.csv')
# Calculate metrics for Ridge Regression
mae_kaggle_test = mean_absolute_error(y_test, y_pred_kaggle_test)
r2_kaggle_test = r2_score(y_test, y_pred_kaggle_test)
rmse_kaggle_test = np.sqrt(mean_squared_error(y_test, y_pred_kaggle_test))

# Print kaggle result
print("Kaggle Results:")
print(f'MAE (Test): {mae_kaggle_test:.2f}')
print(f'R² (Test): {r2_kaggle_test:.4f}')
print(f'RMSE (Test): {rmse_kaggle_test:.2f}')

# #Initializes a linear regression model
# lr= LinearRegression()

# #Fits the linear regression model to the training data
# lr.fit(X_train, y_train)

# #prints the mean cross-validation score for the training data.
# print(cross_val_score(lr, X_train, y_train).mean())

# #Prints the R-squared scores 
# lr.score(X_train, y_train), lr.score(X_test, y_test)

---

### Generate predictions on test data

#generates prediction
prediction_kaggle = lr.predict(X_train)


prediction_kaggle = prediction_kaggle[:len(df_cleaned)]
df_cleaned.loc[:877, 'SalePrice'] = prediction_kaggle[:878]

df_cleaned.head()

submission_df = df_cleaned[['Id', 'SalePrice']]

#Save prediction to CSV

#submission_df.to_csv('AA6prediction.csv', index=False)
